% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_add_entity}
\alias{a_add_entity}
\title{a_add_entity}
\usage{
a_add_entity(tag, id, className = "", parentEntityId = "")
}
\arguments{
\item{tag}{primitive A-Frame entity.}

\item{id}{the id of the new entity to be created}

\item{className}{the class of the new entity to be created}

\item{parentEntityId}{new entity nested within parent via its id}
}
\value{
An object that represents an A-Frame Event.
}
\description{
Add an A-Frame entity from the scene
}
\details{
Add the entity identified by a 'tag' and an 'id' input.
}
