% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{a_asset}
\alias{a_asset}
\title{a_asset}
\usage{
a_asset(...)
}
\arguments{
\item{id}{an an id to be used by the asset item in the asset block. #id will be used to refernce the asset in component configuratuion.}

\item{tag}{the tag text to use in the asset block. Defaults to `a-asset-item`.}

\item{src}{the location of the asset}

\item{inline}{boolean signifying if the asset is to be specified inline with the entity. If true, the containing A-Frame scene does not wait for the asset to load.}
}
\description{
Create an A-Frame asset
}
\details{
This function is used to create an asset that will be loaded before an
A-Frame scene loads in the browser. Assets are typically models, images,
videos, and audio files. The asset object understands there may be two
references to the asset: one in the <a-assets> block and the second refernce
in a component configuratuion. It knows how to render itself for both of
these differnt contexts. It can also be called with `inline=TRUE` which has
the effect of omitting it from the <a-assets> block and inserting a direct
link into the component configuratuion.
}
