% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_control_camera.R
\name{a_pc_control_camera}
\alias{a_pc_control_camera}
\title{a_pc_control_camera}
\usage{
a_pc_control_camera(
  .js_sources = list(),
  easingY = 15,
  acceleration = 100,
  .version_num = "4.1.2",
  ...
)
}
\arguments{
\item{.js_sources}{Defaults to version 4.1.2 of the controls.}

\item{easingY}{Easing rate of movement in the y direction (up).}

\item{acceleration}{Acceleration of movement in forward, backward, left, and
right directions.}

\item{.version_num}{A-Frame extras version number.}

\item{...}{additional components to be added to this entity.}
}
\value{
An entity object describing a camera controlled by keyboard and mouse.
}
\description{
Personal computer movement controls for an A-Frame scene camera
}
\details{
This function returns an entity that will enable the scene camera perspective to be controlled using WASD keys and the mouse.

The entity imports an external js source for controls written by Don McCurdy.
}
