% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_remove_component}
\alias{a_remove_component}
\title{a_remove_component}
\usage{
a_remove_component(id, component)
}
\arguments{
\item{id}{the id of the entity to have component removed}

\item{component}{the name of the component to remove.}
}
\value{
An object that represents an A-Frame event.
}
\description{
Remove a component from an A-Frame entity
}
\details{
Remove `component` from the entity with `id`.
}
