% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_remove_entity_class}
\alias{a_remove_entity_class}
\title{a_remove_entity_class}
\usage{
a_remove_entity_class(className)
}
\arguments{
\item{className}{class of the entity to be removed.}
}
\value{
An object that represents an A-Frame Event.
}
\description{
Remove all A-Frame entities from the scene by class
}
\details{
Remove the entitities identified by 'className' from the scene.
}
