% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{a_update}
\alias{a_update}
\title{a_update}
\usage{
a_update(id, component, attributes, replaces_component = FALSE)
}
\arguments{
\item{id}{the id of the entity that will have the event emitted on it.}

\item{component}{the name of the component to have its attributes updated.}

\item{attributes}{the attributes to be set along with their values.}

\item{replaces_component}{if TRUE, any attributes not specified are set to
their default values. This effectively makes this update a full
replacement of the existing component.}
}
\value{
An object that represents an A-Frame event.
}
\description{
Build an A-Frame Update Message
}
\details{
Construct an A-Frame component attribute update message to be applied to the
component of entity identified by `id`.

Note for setting vector components: Due to a quirk in the A-Frame JS API,
setting vectors for position, rotation, scale etc must be done with a named
list: i.e.

a_update(id = "my_box",
         component = "position",
         attributes = list(x = 1, y = 1, z = 1))
}
