% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_question_scene.R
\name{binary_question_scene}
\alias{binary_question_scene}
\title{Create a binary question scene}
\usage{
binary_question_scene(
  the_question,
  answer_1,
  answer_2,
  img_paths,
  message_height = 1.5
)
}
\arguments{
\item{the_question}{}

\item{answer_1}{}

\item{answer_2}{}

\item{img_paths}{}
}
\value{

}
\description{
Create a binary question scene
}
\examples{
\donttest{
animals <- binary_question_scene("Do you see any koalas in this image?", "Yes", "No", img_paths)
}

}
