% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_2_raster.R
\name{contour_2_raster}
\alias{contour_2_raster}
\title{Convert a shapefile with contours to a raster}
\usage{
contour_2_raster(contour_sf, raster_attribute, n_samples = 10000, nrow, ncol)
}
\arguments{
\item{contour_sf}{an sf object read from a shapefile with contours, eg elevation.}

\item{raster_attribute}{The name of the attribute to be rastered}

\item{n_samples}{the number of contour points to be sampled for raster interpolation}

\item{nrow}{number of rows in the raster grid}

\item{ncol}{number of columns in the raster grid}
}
\value{
a raster
}
\description{
Convert a shapefile with contours to a raster
}
