% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_IP.R
\name{find_IP}
\alias{find_IP}
\title{Find local IP address}
\usage{
find_IP(IP = "IPv4")
}
\arguments{
\item{IP}{IP identity from \code{ipconfig}. Defaults to \code{"IPv4"}}
}
\value{

}
\description{
Find local IP address
}
\examples{
find_IP()
}
