% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Go to next VR scene}
\usage{
go(image_paths, index = NA, question_type = "binary")
}
\arguments{
\item{image_paths}{character string of image paths}

\item{index}{optional input to select a specific image}

\item{question_type}{Question type from "binary" or "multivariable". Defaults to \code{"binary"}}
}
\description{
Go to next VR scene
}
\examples{
\donttest{
image_paths <- c("img1.jpg", "img2.jpg", "img3.jpg")

go(image_paths, 2)
go(image_paths, 3)
}

}
