% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_3d.R
\name{is_3d_point}
\alias{is_3d_point}
\title{is_3d_point}
\usage{
is_3d_point(point)
}
\arguments{
\item{point}{an object}
}
\value{
boolean
}
\description{
Checks if is this object a 3d point
}
\examples{
\donttest{
p1 <- point_3d(1, 0.5, 0.5, 0.5)
is_3d_point(p1) # TRUE
is_3d_point(5) # FALSE
}

}
