% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_question_scene.R
\name{multivariable_question_scene}
\alias{multivariable_question_scene}
\title{Create a multivariable question scene}
\usage{
multivariable_question_scene(
  the_question,
  answer_1,
  answer_2,
  answer_3,
  answer_4,
  img_paths,
  message_height = 1.5
)
}
\arguments{
\item{the_question}{}

\item{answer_1}{}

\item{answer_2}{}

\item{answer_3}{}

\item{answer_4}{}

\item{img_paths}{}

\item{message_height}{}
}
\value{

}
\description{
Create a multivariable question scene
}
\examples{
\donttest{
animals <- multivariable_question_scene("Do you see any of these habitat features in this image? If you do see a feature, click on the box to select it.",
"water", "Jaguar tracks", "Scratch marks", "Dense Vegetation", img_paths)
}

}
