% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_2d.R
\name{point_2d}
\alias{point_2d}
\title{Create new 2d Point instance}
\usage{
point_2d(n, x, y)
}
\arguments{
\item{n}{Integer (uniquely) defining the number of the point to be created}

\item{x}{x position of the point}

\item{y}{y position of the point}
}
\description{
Create new 2d Point instance
}
\examples{
\donttest{
p1 <- point_2d(1, 0.5, 0.5)
}

}
