% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_3d.R
\name{point_3d}
\alias{point_3d}
\title{Create new 3d Point instance}
\usage{
point_3d(n, x, y, z)
}
\arguments{
\item{n}{Integer (uniquely) defining the number of the point to be created}

\item{x}{x position of the point}

\item{y}{y position of the point}

\item{z}{z position of the point}
}
\description{
Create new 3d Point instance
}
\examples{
\donttest{
p1 <- point_3d(1, 0.5, 0.5, 0.5)
}

}
