% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_in_sf.R
\name{point_in_sf}
\alias{point_in_sf}
\title{Determine if a list of X,Y points interest with an sf object.}
\usage{
point_in_sf(x, y, sf_obj)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}

\item{sf_obj}{An sf object, could contian multiple polygons etc}
}
\value{
a logical vector of the same length as the number of points
}
\description{
Determine if a list of X,Y points interest with an sf object.
}
