% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop messages in VR scene}
\usage{
pop(animal_messages, visible = TRUE)
}
\arguments{
\item{animal_messages}{list of a_update() objects}

\item{visible}{optional logical to unpop or repop}
}
\description{
Pop messages in VR scene
}
