% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_random_marker_position_3d.R
\name{random_fixed_3d_marker}
\alias{random_fixed_3d_marker}
\title{Helper function that prints a random marker to be included within a list of fixed images at a default depth (away from the camera). NOTE: change `id` and `isCoral`}
\usage{
random_fixed_3d_marker()
}
\description{
Helper function that prints a random marker to be included within a list of fixed images at a default depth (away from the camera). NOTE: change `id` and `isCoral`
}
\examples{
\donttest{
random_fixed_3d_marker()
[1] "list(id = 1, x = 0.876366487435723, y = -0.00865718001555315, z = -0.481567059643567, isCoral = 0),"
random_fixed_3d_marker()
[1] "list(id = 1, x = 0.729674742517203, y = -0.312136177299759, z = 0.608396069146693, isCoral = 0),"

# training_3d.R
img1Points = list(list(id = 1, x = 0.876366487435723, y = -0.00865718001555315, z = -0.481567059643567, isCoral = 0), list(id = 1, x = 0.729674742517203, y = -0.312136177299759, z = 0.608396069146693, isCoral = 0))
}

}
