% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_randomize_markers.R
\name{randomize_markers}
\alias{randomize_markers}
\title{Randomize the position of the markers}
\usage{
randomize_markers(module = MODULE)
}
\arguments{
\item{module}{"2d" || "3d"}
}
\description{
Randomize the position of the markers
}
\examples{
\donttest{
randomize_markers()
randomize_markers("2d")
randomize_markers("3d")
}

}
