% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Get data from database with API GET request}
\usage{
read(url)
}
\arguments{
\item{url}{url to API end point}
}
\value{

}
\description{
Get data from database with API GET request
}
\examples{
\donttest{
data.df <- read("https://test-api-koala.herokuapp.com/koala")
}
}
