% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_reset_markers_color.R
\name{reset_markers_color}
\alias{reset_markers_color}
\title{Reset the color of all markers}
\usage{
reset_markers_color(
  number_of_markers = NUMBER_OF_MARKERS,
  default_marker_color = COLOR_MARKER
)
}
\arguments{
\item{number_of_markers}{Integer to indicate the number of markers to reset the color of}

\item{default_marker_color}{Character string of the HEX color to reset the markers back to}
}
\description{
Reset the color of all markers
}
\examples{
\donttest{
NUMBER_OF_MARKERS <- 15
COLOR_MARKER <- "#FFFFFF"
reset_markers_color()

reset_markers_color(15, "#FFFFFF")
}

}
