% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_set_colors.R
\name{set_colors}
\alias{set_colors}
\title{Set the colors of the user interface entities}
\usage{
set_colors(
  marker = "#FFFFFF",
  coral = "#FF95BC",
  not_coral = "#B8B27B",
  text = "#000000",
  plane = "#FFFFFF",
  check_correct = "#00FF00",
  check_incorrect = "#FF0000",
  evaluation_selection = "#00FF00",
  cursor = "#FF0000"
)
}
\arguments{
\item{marker}{Character string of the HEX color}

\item{coral}{Character string of the HEX color}

\item{not_coral}{Character string of the HEX color}

\item{text}{Character string of the HEX color}

\item{plane}{Character string of the HEX color}

\item{check_correct}{Character string of the HEX color}

\item{check_incorrect}{Character string of the HEX color}

\item{evaluation_selection}{Character string of the HEX color}

\item{cursor}{Character string of the HEX color}
}
\description{
Set the colors of the user interface entities
}
\examples{
\donttest{
set_colors()

set_colors("#FFFFFF")

set_colors("#FFFFFF", "#FF95BC", "#B8B27B", "#000000", "#FFFFFF", "#00FF00", "#FF0000", "#00FF00", "#FF0000")
}

}
