% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_set_go_to_context.R
\name{set_go_to_context}
\alias{set_go_to_context}
\title{Set global variables needed for go_to() and check() to assess which image is being displayed}
\usage{
set_go_to_context(image_paths_and_points)
}
\arguments{
\item{image_paths_and_points}{List of Lists containing the img and the points for the annotation markers}
}
\description{
Set global variables needed for go_to() and check() to assess which image is being displayed
}
\examples{
\donttest{
img_paths_and_points <- list(
  list(img = img_paths[1], img_points = img1Points[1]),
  list(img = img_paths[2], img_points = img1Points[2]),
  list(img = img_paths[3], img_points = img1Points[3]),
  list(img = img_paths[4], img_points = img1Points[1])
) 
 
set_go_to_context(img_paths_and_points)                                  
}

}
