% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_set_user.R
\name{set_user}
\alias{set_user}
\title{Set the full name of the user globally}
\usage{
set_user(full_name)
}
\arguments{
\item{full_name}{String ideally in the form 'Firstname-LastName' e.g. 'John-Doe'}
}
\description{
Set the full name of the user globally
}
\examples{
\donttest{
set_user("Jon-Peppinck")
}

}
