% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_generate_markers.R
\name{shared_generate_markers}
\alias{shared_generate_markers}
\title{Generates the entities needed for the specified number of markers for the module}
\usage{
shared_generate_markers(module = MODULE, module_type = MODULE_TYPE)
}
\arguments{
\item{module}{String to indicate if "2d" or "3d"}
}
\description{
Generates the entities needed for the specified number of markers for the module
}
\examples{
\donttest{
shared_generate_markers("2d")
shared_generate_markers("3d")
}

}
