% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_model.R
\name{a_json_model}
\alias{a_json_model}
\title{a_json_model}
\usage{
a_json_model(
  src,
  .version_num = "4.1.2",
  mesh_smooth = FALSE,
  .js_sources = list(),
  ...
)
}
\arguments{
\item{src}{an a_asset describing a JSON file.}

\item{.version_num}{A-Frame extras version number.}

\item{mesh_smooth}{if true, compute normals in browser to give the model a
smoothed appearance.}

\item{.js_sources}{defaults to version 4.1.2 of the JSON model loader, supply
another version using a list here. e.g. `js_sources` =
list("https://cdn.rawgit.com/donmccurdy/aframe-extras/<path_to_js_file>")}

\item{...}{other components to be added to the JSON model. Passed to a_entity.}
}
\value{
An entity object describing a JSON model.
}
\description{
This function returns an entity that will render a 3D model expressed in
threejs JSON format. See the format specification here:
https://github.com/mrdoob/three.js/wiki/JSON-Model-format-3. The mandatory
`src` argument should be an object returned by `a_asset()`.
}
\details{
The entity imports the an extenal js source for loading JSON models written by Don McCurdy.
}
