% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Go to next VR scene}
\usage{
go(image_paths, index = NA, question_type = "binary")
}
\arguments{
\item{image_paths}{character string of image paths}

\item{index}{optional input to select a specific image}
}
\description{
Go to next VR scene
}
\examples{
\donttest{

image_paths <- c("img1.jpg", "img2.jpg", "img3.jpg")

setup_scene <- list(
a_update(id = "canvas3d",
         component = "material",
         attributes = list(src = paste0("#",next_image))),
a_update(id = "canvas3d",
         component = "src",
        attributes = paste0("#",next_image)),
a_update(id = "canvas3d",
         component = "rotation",
         attributes = context_rotations),
a_update(id = "canvas3d",
         component = "class",
         attributes = image_paths),
a_update(id = "yesPlane",
         component = "color",
         attributes = white),
a_update(id = "noPlane",
         component = "color",
         attributes = white))

go(image_paths, setup_scene)

}

}
