% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Go to next VR scene}
\usage{
go(image_paths, setup_scene, index = NA)
}
\arguments{
\item{image_paths}{character string of image paths}

\item{setup_scene}{aframe entities for scene setup}

\item{index}{optional input to select a specific image}
}
\description{
Go to next VR scene
}
